separate file for FloatingLocalVideo
