+++ packages/react-components/src/components/VideoGallery/FloatingLocalVideoLayout.tsx
@@ -1,25 +1,21 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ContextualMenu, IDragOptions, LayerHost, mergeStyles, Stack } from '@fluentui/react';
+import { LayerHost, mergeStyles, Stack } from '@fluentui/react';
 import { useId } from '@fluentui/react-hooks';
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useTheme } from '../../theming';
 import { VideoGalleryRemoteParticipant } from '../../types';
 import { GridLayout } from '../GridLayout';
-import { _ICoordinates, _ModalClone } from '../ModalClone/ModalClone';
 import {
-  floatingLocalVideoModalStyle,
-  LARGE_FLOATING_MODAL_SIZE_PX,
   layerHostStyle,
   localVideoTileContainerStyle,
-  localVideoTileOuterPaddingPX,
   localVideoTileWithControlsContainerStyle,
-  LOCAL_VIDEO_TILE_ZINDEX,
-  SMALL_FLOATING_MODAL_SIZE_PX
+  LOCAL_VIDEO_TILE_ZINDEX
 } from '../styles/VideoGallery.styles';
 import { isNarrowWidth } from '../utils/responsive';
 import { VideoGalleryStyles } from '../VideoGallery';
+import { FloatingLocalVideo } from './FloatingLocalVideo';
 import { rootLayoutStyle } from './styles/DefaultLayout.styles';
 import { VideoGalleryResponsiveHorizontalGallery } from './VideoGalleryResponsiveHorizontalGallery';
 import { useFloatingLocalVideoLayout } from './videoGalleryUtils';
@@ -62,19 +58,6 @@ export interface FloatingLocalVideoLayoutProps {
    */
   parentHeight?: number;
 }
-
-const DRAG_OPTIONS: IDragOptions = {
-  moveMenuItemText: 'Move',
-  closeMenuItemText: 'Close',
-  menu: ContextualMenu,
-  keepInBounds: true
-};
-
-// Manually override the max position used to keep the modal in the bounds of its container.
-// This is a workaround for: https://github.com/microsoft/fluentui/issues/20122
-// Because our modal starts in the bottom right corner, we can say that this is the max (i.e. rightmost and bottomost)
-// position the modal can be dragged to.
-const modalMaxDragPosition = { x: localVideoTileOuterPaddingPX, y: localVideoTileOuterPaddingPX };
 
 /**
  * FloatingLocalVideoLayout displays remote participants and a screen sharing component in
@@ -182,48 +165,5 @@ export const FloatingLocalVideoLayout = (props: FloatingLocalVideoLayoutProps): 
         <LayerHost id={layerHostId} className={mergeStyles(layerHostStyle)} />
       </Stack>
     </Stack>
-  );
-};
-
-const FloatingLocalVideo = (props: {
-  localVideoComponent: JSX.Element;
-  layerHostId: string;
-  isNarrow?: boolean;
-  parentWidth?: number;
-  parentHeight?: number;
-}): JSX.Element => {
-  const { localVideoComponent, layerHostId, isNarrow, parentWidth, parentHeight } = props;
-
-  const theme = useTheme();
-
-  const modalWidth = isNarrow ? SMALL_FLOATING_MODAL_SIZE_PX.width : LARGE_FLOATING_MODAL_SIZE_PX.width;
-  const modalHeight = isNarrow ? SMALL_FLOATING_MODAL_SIZE_PX.height : LARGE_FLOATING_MODAL_SIZE_PX.height;
-  // The minimum drag position is the top left of the video gallery. i.e. the modal (PiP) should not be able
-  // to be dragged offscreen and these are the top and left bounds of that calculation.
-  const modalMinDragPosition: _ICoordinates | undefined = useMemo(
-    () =>
-      parentWidth && parentHeight
-        ? {
-            // We use -parentWidth/Height because our modal is positioned to start in the bottom right,
-            // hence (0,0) is the bottom right of the video gallery.
-            x: -parentWidth + modalWidth + localVideoTileOuterPaddingPX,
-            y: -parentHeight + modalHeight + localVideoTileOuterPaddingPX
-          }
-        : undefined,
-    [parentHeight, parentWidth, modalHeight, modalWidth]
-  );
-
-  return (
-    <_ModalClone
-      isOpen={true}
-      isModeless={true}
-      dragOptions={DRAG_OPTIONS}
-      styles={floatingLocalVideoModalStyle(theme, isNarrow)}
-      layerProps={{ hostId: layerHostId }}
-      maxDragPosition={modalMaxDragPosition}
-      minDragPosition={modalMinDragPosition}
-    >
-      {localVideoComponent}
-    </_ModalClone>
   );
 };
